test(mocks): refactor mock classes into their own modules
