+++ test/unit/mocks/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = {
+  Client: require('./client'),
+  Connection: require('./connection'),
+  Session: require('./session'),
+  SenderLink: require('./sender_link'),
+  ReceiverLink: require('./receiver_link')
+};
