+++ test/unit/test_amqpclient.js
@@ -11,7 +11,8 @@ var debug = require('debug')('amqp10-test_amqpclient'),
     Connection = require('../../lib/connection'),
     Session = require('../../lib/session'),
     Link = require('../../lib/link'),
-    Sender = require('../../lib/sender_link'),
+
+    Mock = require('./mocks'),
 
     u = require('../../lib/utilities'),
     tu = require('./testing_utils');
@@ -21,134 +22,12 @@ chai.config.includeStack = true; // turn on stack trace
 
 var mock_uri = 'amqp://localhost/';
 
-function MockConnection() {
-  this._created = 0;
-}
-
-util.inherits(MockConnection, EventEmitter);
-
-MockConnection.prototype.open = function(addr, sasl) {
-  this.emit('open-called', this, addr, sasl);
-};
-
-MockConnection.prototype.close = function() {
-  this.emit('close-called', this);
-};
-
-function MockSession(conn) {
-  this._created = 0;
-  this.connection = conn;
-  this._mockLinks = {};
-}
-
-util.inherits(MockSession, EventEmitter);
-
-MockSession.prototype.begin = function(policy) {
-  this.emit('begin-called', this, policy);
-};
-
-MockSession.prototype.attachLink = function(policy) {
-  var link = this._mockLinks[policy.options.name];
-  expect(link).to.exist;
-  link._created++;
-  link._clearState();
-  this.emit('attachLink-called', this, policy, link);
-};
-
-MockSession.prototype._addMockLink = function(link) {
-  this._mockLinks[link.name] = link;
-};
-
-function MockLink(session, options) {
-  this._created = 0;
-  this.session = session;
-  this.options = options;
-  this._clearState();
-}
-
-util.inherits(MockLink, EventEmitter);
-
-MockLink.prototype._clearState = function() {
-  this.name = this.options.name;
-  this.isSender = this.options.isSender || false;
-  this.capacity = this.options.capacity || 0;
-  this.messages = [];
-  this.curId = 0;
-};
-
-function MockSenderLink(session, options) {
-  this._created = 0;
-  this.session = session;
-  this.options = options;
-  this._clearState();
-}
-
-util.inherits(MockSenderLink, Sender);
-
-MockSenderLink.prototype.canSend = function() {
-  this.emit('canSend-called', this);
-  return this.capacity > 0;
-};
-
-MockSenderLink.prototype._sendMessage = function(msg, options) {
-  var self = this;
-  self.curId++;
-  self.messages.push({ id: self.curId, message: msg.body[0], options: options });
-  self.emit('sendMessage-called', self, self.curId, msg, options);
-  return self.curId;
-};
-
-MockSenderLink.prototype._clearState = function() {
-  this.name = this.options.name;
-  this.isSender = this.options.isSender || false;
-  this.capacity = this.options.capacity || 0;
-  this.messages = [];
-  this.curId = 0;
-};
-
-function MakeMockClient(c, s) {
-  var client = new AMQPClient();
-  client._newConnection = function() {
-    client._clearState();
-    c._created++;
-    return c;
-  };
-
-  client._newSession = function(conn) {
-    for (var lname in s._mockLinks) {
-      var l = s._mockLinks[lname];
-      [Link.MessageReceived, Link.ErrorReceived, Link.CreditChange, Link.Detached].forEach(l.removeAllListeners);
-    }
-    expect(c).to.equal(conn);
-    s._created++;
-    return s;
-  };
-
-  client._clearState = function() {
-    var connEvts = [Connection.Connected, Connection.Disconnected];
-    var sessEvts = [Session.Mapped, Session.Unmapped, Session.ErrorReceived, Session.LinkAttached, Session.LinkDetached, Session.DispositionReceived];
-    var idx, e;
-    for (idx in connEvts) {
-      e = connEvts[idx];
-      c.removeAllListeners(e);
-      expect(EventEmitter.listenerCount(c, e)).to.eql(0);
-    }
-    for (idx in sessEvts) {
-      e = sessEvts[idx];
-      s.removeAllListeners(e);
-      expect(EventEmitter.listenerCount(s, e)).to.eql(0);
-    }
-  };
-
-  return client;
-}
-
 describe('AMQPClient', function() {
   describe('#connect()', function() {
     it('should set up connection and session', function() {
-      var c = new MockConnection();
-      var s = new MockSession(c);
-      var client = new MakeMockClient(c, s);
+      var c = new Mock.Connection();
+      var s = new Mock.Session(c);
+      var client = new Mock.Client(c, s);
       var called = {open: 0, begin: 0};
       c.on('open-called', function(_c, _addr, _sasl) {
         expect(_addr).to.eql(u.parseAddress(mock_uri));
@@ -173,23 +52,23 @@ describe('AMQPClient', function() {
 
   describe('#send()', function() {
     it('should throw an error if not connected', function () {
-      var client = new MakeMockClient();
+      var client = new Mock.Client();
       expect(function () {
         client.send();
       }).to.throw(Error);
     });
 
     it('should wait for capacity before sending', function () {
-      var c = new MockConnection();
-      var s = new MockSession(c);
-      var l = new MockSenderLink(s, {
+      var c = new Mock.Connection();
+      var s = new Mock.Session(c);
+      var l = new Mock.SenderLink(s, {
         name: 'queue_TX',
         isSender: true,
         capacity: 0
       });
 
       s._addMockLink(l);
-      var client = new MakeMockClient(c, s);
+      var client = new Mock.Client(c, s);
       var queue = 'queue';
       var called = {open: 0, begin: 0, attachLink: 0, canSend: 0, sendMessage: 0};
       c.on('open-called', function (_c, _addr, _sasl) {
@@ -238,7 +117,7 @@ describe('AMQPClient', function() {
           return client.createSender(queue);
         })
         .then(function (sender) {
-          sender.send({my: 'message'});
+          return sender.send({ my: 'message' });
         })
         .then(function () {
           expect(c._created).to.eql(1);
@@ -247,7 +126,7 @@ describe('AMQPClient', function() {
           expect(called.open).to.eql(1);
           expect(called.begin).to.eql(1);
           expect(called.attachLink).to.eql(1);
-          expect(called.canSend).to.eql(2);
+          expect(called.canSend).to.eql(1);
           expect(called.sendMessage).to.eql(1);
           expect(l.messages).to.not.be.empty;
           expect(l.messages[0].message).to.eql({my: 'message'});
@@ -257,15 +136,14 @@ describe('AMQPClient', function() {
 
   describe('#receive()', function() {
     it('should throw an error if not connected', function() {
-      var client = new MakeMockClient();
+      var client = new Mock.Client();
       expect(function() { client.createReceiver(); }).to.throw(Error);
     });
 
-
     function DelayedAttachMockSession(conn) {
-      MockSession.call(this);
+      Mock.Session.call(this);
     }
-    util.inherits(DelayedAttachMockSession, MockSession);
+    util.inherits(DelayedAttachMockSession, Mock.Session);
 
     DelayedAttachMockSession.prototype.attachLink = function(policy) {
       var link = this._mockLinks[policy.options.name];
@@ -281,16 +159,16 @@ describe('AMQPClient', function() {
     };
 
     it('should return cached receiver links upon multiple createReceive calls', function() {
-      var c = new MockConnection();
+      var c = new Mock.Connection();
       var s = new DelayedAttachMockSession(c);
-      var l = new MockLink(s, {
+      var l = new Mock.ReceiverLink(s, {
         name: 'queue_RX',
         isSender: false,
         capacity: 100
       });
 
       s._addMockLink(l);
-      var client = new MakeMockClient(c, s);
+      var client = new Mock.Client(c, s);
       var queue = 'queue';
       var called = { open: 0, begin: 0, attachLink: 0 };
       c.on('open-called', function(_c, _addr, _sasl) {
@@ -340,16 +218,16 @@ describe('AMQPClient', function() {
 
 
     it('should create connection, session, and link on receive with full address', function() {
-      var c = new MockConnection();
-      var s = new MockSession(c);
-      var l = new MockLink(s, {
+      var c = new Mock.Connection();
+      var s = new Mock.Session(c);
+      var l = new Mock.ReceiverLink(s, {
         name: 'queue_RX',
         isSender: false,
         capacity: 100
       });
 
       s._addMockLink(l);
-      var client = new MakeMockClient(c, s);
+      var client = new Mock.Client(c, s);
       var queue = 'queue';
       var called = { open: 0, begin: 0, attachLink: 0 };
       c.on('open-called', function(_c, _addr, _sasl) {
@@ -387,15 +265,15 @@ describe('AMQPClient', function() {
     });
 
     it('should only create a single connection, session, multiple links for multiple receives', function() {
-      var c = new MockConnection();
-      var s = new MockSession(c);
-      var l1 = new MockLink(s, {
+      var c = new Mock.Connection();
+      var s = new Mock.Session(c);
+      var l1 = new Mock.ReceiverLink(s, {
         name: 'queue1_RX',
         isSender: false,
         capacity: 100
       });
 
-      var l2 = new MockLink(s, {
+      var l2 = new Mock.ReceiverLink(s, {
         name: 'queue2_RX',
         isSender: false,
         capacity: 100
@@ -404,7 +282,7 @@ describe('AMQPClient', function() {
       s._addMockLink(l1);
       s._addMockLink(l2);
 
-      var client = new MakeMockClient(c, s);
+      var client = new Mock.Client(c, s);
       var called = { open: 0, begin: 0, attachLink: 0 };
       c.on('open-called', function(_c, _addr, _sasl) {
         expect(_addr).to.eql(u.parseAddress(mock_uri));
@@ -444,16 +322,16 @@ describe('AMQPClient', function() {
     });
 
     it('should re-establish receive link on detach, automatically', function() {
-      var c = new MockConnection();
-      var s = new MockSession(c);
-      var l = new MockLink(s, {
+      var c = new Mock.Connection();
+      var s = new Mock.Session(c);
+      var l = new Mock.ReceiverLink(s, {
         name: 'queue_RX',
         isSender: false,
         capacity: 100
       });
 
       s._addMockLink(l);
-      var client = new MakeMockClient(c, s);
+      var client = new Mock.Client(c, s);
       var queue = 'queue';
       var called = { open: 0, begin: 0, attachLink: 0 };
       c.on('open-called', function(_c, _addr, _sasl) {
@@ -499,16 +377,16 @@ describe('AMQPClient', function() {
     });
 
     it('should re-establish connection on disconnect, automatically', function() {
-      var c = new MockConnection();
-      var s = new MockSession(c);
-      var l = new MockLink(s, {
+      var c = new Mock.Connection();
+      var s = new Mock.Session(c);
+      var l = new Mock.ReceiverLink(s, {
         name: 'queue_RX',
         isSender: false,
         capacity: 100
       });
 
       s._addMockLink(l);
-      var client = new MakeMockClient(c, s);
+      var client = new Mock.Client(c, s);
       var queue = 'queue';
       var called = { open: 0, begin: 0, attachLink: 0 };
       c.on('open-called', function(_c, _addr, _sasl) {
